Split out parts of the backend
