+++ src/main/deltachat/chats.js
@@ -1,0 +1,3 @@
+
+module.exports = function() {
+}
